Contact Management System
The user can choose to:
1. Manage their contact list, where they can add, delete and edit their contacts.
2. View their contact list, where they can see view all of their
   contacts sorted based on the filter (by name, phone number or email) they selected.
3. Search contact list, where they can search a contact through the
   filter (by name, phone number or email) they selected.

Implementation of mergesort, quicksort and binary search in the system.

MergeSort is used to sort the contact list before they are displayed in
the view contact list option, based on the filter the user selected,
this is used in the view feature of the program.

QuickSort is used to sort the the contact list based again on the filter
the user selected before proceding with binary search.

Binary search is responsible to find if the user search input is within
their contact list again based on the filter the user selected.
Both quicksort and binary search is used on searching feature of the
program.