Keyword searching
I added a new option in the searching feature of the program it search
the contacts name, phone number, and email, for the string that is
entered by the user if that inputted string is a substring of any of the fields
of the contact object, the program will display the contact.
The searching is done through linear search algorithm.

I also create a custom type using the "using" keyword, that represents a
function pointer since i find it annoying since the function pointer
parameter is really long, so by just using the new custom type i can
shorter it at least.